+++ example/android/app/src/main/kotlin/com/adriankohls/bubble_overlay_example/MainActivity.kt
@@ -1,6 +1,0 @@
-package com.adriankohls.bubble_overlay_example
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}