+++ CHANGELOG.md
@@ -8,6 +8,18 @@
 
 * Initial release.
 
-## 0.0.1+7
+## 0.0.2
 
-* Removed close bubble on tapping on bubble, close bubble overlay ony tapping on X
+* Removed close bubble on tapping on bubble, close bubble overlay only tapping on X
+* Added top and bottom text to bubble
+* Fixed title, middle and bottom text size
+
+## 0.0.3
+
+* Fixed close button layout
+* Fixed bubble layouts
+* Renamed title to topText
+* Fixed bubble text sizes
+* Added Timer callback, for custom inside bubble functions
+* Fixed some access to object
+